+++ Changelog.md
@@ -4,9 +4,11 @@ The changes for each individual (beta) release can be seen [here](https://github
 
 ## 2.2.6.0
 
+- CORS updates to support all HTTP verbs
 - Added `api/time` to get the current time, can be used to sync with other systems
 - Added `api/application-start` to get the application start date
 - Added `api/version/nina` to get the version of NINA
+- Added `api/application/logs` to get the last N log entries
 - Added `sequence/load` to load sequences into the advanced sequence
   - Use GET to load a sequence from the default sequence folder (you can use the `sequence/list-available` endpoint to get the available sequences)
   - Use POST to load a sequence from a json string, supplied by the client in the request body